feat: message components refac
