+++ src/lib/components/chat/Messages/Name.svelte
@@ -1,0 +1,3 @@
+<div class=" self-center font-bold mb-0.5 capitalize">
+	<slot />
+</div>
